requirement.txt
# Core
numpy
pandas
scikit-learn
scipy
joblib
tqdm

# UI
streamlit

# Chemistry
rdkit-pypi

# Viz
matplotlib
